frames: new compressed format for arrays of ints
The idea is to store the first integer (2 bytes), and for the integers
after the first one store the difference only (1 byte).

For instance if we have the array [330, 328, 335] then store the values:
330 (2 bytes), -2 (1 byte), +7 (1 byte).

So, if we have for instance an array of 30 integers, before the change
it would take: sensor type (1 byte), number of values (1 byte), 30
values (2 bytes per value). That makes a total of 62 bytes. With the
new system it only takes 33 bytes.

Now, if the difference doesn't fit in one byte what we do is intrduce
a marker byte (value -128) and then 2 bytes with the full value. For
instance the array [330, 1360, 335] would be stored as: 330, -128, 1360,
-128, 335.

In this worst case scenario the new system is less efficient than the
old system. But if we have values like there's probably something wrong
(at least with a string of DS18B20 sensors, the use case this has been
developed for).

In the case of a string of DS18B20 sensors, the temperature difference
to trigger this special case is 8 degrees, which I think should not
happen in real conditions.